Finalizes scenes 3-7.
Unity is crashing every couple of minutes for me. It's triggered by
viewing the "wrong" part of large geometry in the scene viewer. I
can't even open VideoScene2 anymore without Unity freezing up. Will
try a reinstall at some point, but this is a blocker for getting
the opening scene to work.